+++ shared/js/ui/frames/side/PrivateConversationManager.ts
@@ -2,6 +2,7 @@ import {ClientEntry} from "tc-shared/ui/client";
 import {ConnectionHandler, ConnectionState} from "tc-shared/ConnectionHandler";
 import {EventHandler, Registry} from "tc-shared/events";
 import {
+    PrivateConversationManagerEvents,
     PrivateConversationInfo,
     PrivateConversationUIEvents
 } from "tc-shared/ui/frames/side/PrivateConversationDefinitions";
@@ -252,6 +253,9 @@ export class PrivateConversation extends AbstractChat<PrivateConversationUIEvent
         /* TODO: Move this somehow to the client itself? */
         if(this.activeClient instanceof ClientEntry)
             this.activeClient.setUnread(timestamp !== undefined);
+
+        /* TODO: Eliminate this cross reference? */
+        this.connection.side_bar.info_frame().update_chat_counter();
     }
 
     protected canClientAccessChat(): boolean {
@@ -311,6 +315,7 @@ export class PrivateConversation extends AbstractChat<PrivateConversationUIEvent
 }
 
 export class PrivateConversationManager extends AbstractChatManager<PrivateConversationUIEvents> {
+    public readonly events: Registry<PrivateConversationManagerEvents>;
     public readonly htmlTag: HTMLDivElement;
     public readonly connection: ConnectionHandler;
 
@@ -322,6 +327,7 @@ export class PrivateConversationManager extends AbstractChatManager<PrivateConve
     constructor(connection: ConnectionHandler) {
         super();
         this.connection = connection;
+        this.events = new Registry<PrivateConversationManagerEvents>();
 
         this.htmlTag = document.createElement("div");
         this.htmlTag.style.display = "flex";